chore(ai-research): colocate ML mirror anonymizer
