+++ nodejs/src/ingestion/pipelines/sessionreplay/ml-mirror/parse-and-anonymize-step.test.ts
@@ -4,17 +4,17 @@ import { register } from 'prom-client'
 import { gzip } from 'zlib'
 
 import { PipelineResultType } from '~/ingestion/framework/results'
+import { imageRef } from '~/ingestion/pipelines/sessionreplay/ml-mirror-image-scrub/content-ref'
+import { SessionReplayHeaders } from '~/ingestion/pipelines/sessionreplay/pipeline-types'
 
-import { imageRef } from './ml-mirror-image-scrub/content-ref'
+import { createParseAndAnonymizeMessageStep } from './parse-and-anonymize-step'
 import {
     PSEUDONYM_IMAGE_CONTENT_KEY,
     PSEUDONYM_IMAGE_URL_GLOBAL_VALUE,
     PSEUDONYM_IMAGE_URL_KEY,
     PSEUDONYM_TEAM,
     pseudonymize,
-} from './ml-mirror/pseudonymize'
-import { createParseAndAnonymizeMessageStep } from './parse-and-anonymize-step'
-import { SessionReplayHeaders } from './pipeline-types'
+} from './pseudonymize'
 
 const compressWithGzip = promisify(gzip)
 const IMAGE_SOURCE_METRIC = 'recording_blob_ingestion_v2_ml_image_references_by_property'
